Conflicts:
	amp_conf/htdocs/admin/libraries/media/Media/Media.php
	amp_conf/htdocs/admin/views/footer_content.php
	amp_conf/htdocs/admin/views/header.php

+++ amp_conf/htdocs/admin/libraries/Builtin/UpdateManager.php
@@ -12,6 +12,16 @@ class UpdateManager {
 		$this->machine_id = $settings['system_ident'];
 	}
 
+	private function isCloudDeployment() {
+		if($this->freepbx->Modules->checkStatus('sysadmin')) {
+			if (method_exists($this->freepbx->Sysadmin(),"isCloudDeployment")) {
+				if($this->freepbx->Sysadmin->isCloudDeployment()){
+					return true;
+				}
+			}
+		}
+		return false;
+	}
 	/**
 	 * Handle an Ajax request from the Updates Page
 	 *
@@ -291,6 +301,9 @@ class UpdateManager {
 	 * @param int $priority
 	 */
 	public function sendEmail($tag, $subject, $message, $priority = 4, $force = false) {
+		if($this->isCloudDeployment()){
+			return;
+		}
 		$settings = $this->getCurrentUpdateSettings(false); // Don't html encode the output
 		$to = $settings['notification_emails'];
 
@@ -366,6 +379,9 @@ class UpdateManager {
 	}
 
 	public function securityEmail() {
+		if($this->isCloudDeployment()){
+			return;
+		}
 		$security = $this->freepbx->Notifications->list_security();
 		if (!$security) {
 			return;
@@ -390,6 +406,9 @@ class UpdateManager {
 	}
 
 	public function updateEmail($willupdate = false) {
+		if($this->isCloudDeployment()){
+			return;
+		}
 		$updates = $this->freepbx->Notifications->list_update();
 		if (!$updates) {
 			return;
@@ -417,6 +436,9 @@ class UpdateManager {
 	}
 
 	public function unsignedEmail() {
+		if($this->isCloudDeployment()){
+			return;
+		}
 		if($this->getCurrentUpdateSettings(false)['unsigned_module_emails'] === 'disabled'){
 			$this->freepbx->Notifications->delete("freepbx", "SIGEMAIL");
 			return true;
